requirements.txt
scikit-learn
hmmlearn
polygon-api-client
pandas
yfinance
matplotlib
numpy
